Various security tests do not conform to guidelines
Problem
-------

Various unit tests in tests/security do not conform to guidelines.

Solution
--------

Move and rename according to the guidelines.
