--- tests/security/RTPS/.gitignore
@@ -1,1 +1,0 @@
-/Security_RTPS
