--- tests/security/RTPS/Main.cpp
@@ -1,9 +1,0 @@
-
-#include "gtest/gtest.h"
-
-
-int main(int argc, char* argv[])
-{
-  ::testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}
