--- tests/security/RTPS/RTPS.mpc
@@ -1,10 +1,0 @@
-project (Security*): dcpsexe, dcps_test, dcps_rtps, googletest {
-  requires += no_opendds_safety_profile
-  avoids += no_opendds_security
-  exename = *
-
-  Source_Files {
-    Main.cpp
-    ParameterListConverterTest.cpp
-  }
-}
