--- tests/security/RTPS/run_test.pl
@@ -1,15 +1,0 @@
-eval '(exit $?0)' && eval 'exec perl -S $0 ${1+"$@"}'
-    & eval 'exec perl -S $0 $argv:q'
-    if 0;
-
-use Env (DDS_ROOT);
-use lib "$DDS_ROOT/bin";
-use Env (ACE_ROOT);
-use lib "$ACE_ROOT/bin";
-use PerlDDS::Run_Test;
-use strict;
-
-my $test = new PerlDDS::TestFramework();
-$test->process('test', 'Security_RTPS');
-$test->start_process('test');
-exit $test->finish(60);
