--- tests/security/SSL/.gitignore
@@ -1,2 +1,0 @@
-/Main
-/Security_SSL
