--- tests/security/SSL/SSL.mpc
@@ -1,4 +1,0 @@
-project(Security*): dcpsexe, dcps_test, opendds_security, googletest {
-  requires += no_opendds_safety_profile
-  exename = *
-}
