--- tests/security/framework/.gitignore
@@ -1,1 +1,0 @@
-/OpenDDS_SecurityConfigTest
