--- tests/security/framework/SecurityFrameworkTests.mpc
@@ -1,9 +1,0 @@
-project(SecurityConfigTest): dcpsexe, dcps_test, googlemock {
-  requires += no_opendds_safety_profile
-  exename = OpenDDS_SecurityConfigTest
-  avoids += no_opendds_security
-
-  Source_Files {
-    SecurityConfigTest.cpp
-  }
-}
