--- tests/security/unit_tests/.gitignore
@@ -1,4 +1,0 @@
-/OpenDDS_AuthenticationTest
-/OpenDDS_CryptoKeyFactoryTest
-/OpenDDS_CryptoKeyExchangeTest
-/OpenDDS_CryptoTransformTest
