--- tests/security/unit_tests/security_unit_tests.mpc
@@ -1,35 +1,0 @@
-project(AuthenticationTest): dcpsexe, dcps_test, opendds_security, googletest {
-  requires += no_opendds_safety_profile
-  exename = OpenDDS_AuthenticationTest
-
-  Source_Files {
-    ./Authentication
-  }
-}
-
-project(CryptoKeyFactoryTest): dcpsexe, dcps_test, opendds_security, googletest {
-  requires += no_opendds_safety_profile
-  exename = OpenDDS_CryptoKeyFactoryTest
-
-  Source_Files {
-    CryptoKeyFactoryTest.cpp
-  }
-}
-
-project(CryptoKeyExchangeTest): dcpsexe, dcps_test, opendds_security, googletest {
-  requires += no_opendds_safety_profile
-  exename = OpenDDS_CryptoKeyExchangeTest
-
-  Source_Files {
-    CryptoKeyExchangeTest.cpp
-  }
-}
-
-project(CryptoTransformTest): dcpsexe, dcps_test, opendds_security, googletest {
-  requires += no_opendds_safety_profile
-  exename = OpenDDS_CryptoTransformTest
-
-  Source_Files {
-    CryptoTransformTest.cpp
-  }
-}
