+++ tests/unit-tests/dds/DCPS/RTPS/UnitTests.mpc
@@ -6,8 +6,10 @@ project(*BaseMessageUtils): dcpsexe, dcps_test, dcps_rtps_udp {
   }
 }
 
-project(*ParameterListConverter): dcps_rtpsexe, dcps_test {
+project(*ParameterListConverter): dcps_rtpsexe, dcps_test, googletest {
   exename   = *
+  //  requires += no_opendds_safety_profile
+  //  avoids += no_opendds_security
 
   Source_Files {
     ParameterListConverter.cpp
