+++ tests/unit-tests/dds/DCPS/security/Authentication/UnitTests.mpc
@@ -1,0 +1,8 @@
+project(*LocalAuthCredentialData): dcpsexe, dcps_test, opendds_security, googlemock {
+  requires += no_opendds_safety_profile
+  exename = *
+
+  Source_Files {
+    LocalAuthCredentialData.cpp
+  }
+}
