+++ tests/unit-tests/dds/DCPS/security/AuthenticationBuiltInImpl.cpp
@@ -116,11 +116,11 @@ struct AuthenticationTest : public ::testing::Test
     Property_t idca, pkey, pass, idcert, cperms;
 
     idca.name = "dds.sec.auth.identity_ca";
-    idca.value = "file:../certs/identity/identity_ca_cert.pem";
+    idca.value = "file:./security/certs/identity/identity_ca_cert.pem";
     idca.propagate = false;
 
     pkey.name = "dds.sec.auth.private_key";
-    pkey.value = "file:../certs/identity/test_participant_01_private_key.pem";
+    pkey.value = "file:./security/certs/identity/test_participant_01_private_key.pem";
     pkey.propagate = false;
 
     pass.name = "dds.sec.auth.password";
@@ -128,11 +128,11 @@ struct AuthenticationTest : public ::testing::Test
     pass.propagate = false;
 
     idcert.name = "dds.sec.auth.identity_certificate";
-    idcert.value = "file:../certs/identity/test_participant_01_cert.pem";
+    idcert.value = "file:./security/certs/identity/test_participant_01_cert.pem";
     idcert.propagate = false;
 
     cperms.name = "dds.sec.access.permissions";
-    cperms.value = "file:../permissions/permissions_test_participant_01_JoinDomain_signed.p7s";
+    cperms.value = "file:./security/permissions/permissions_test_participant_01_JoinDomain_signed.p7s";
     cperms.propagate = false;
 
     mp1.add_property(idca);
@@ -151,11 +151,11 @@ struct AuthenticationTest : public ::testing::Test
     Property_t idca, pkey, pass, idcert, cperms;
 
     idca.name = "dds.sec.auth.identity_ca";
-    idca.value = "file:../certs/identity/identity_ca_cert.pem";
+    idca.value = "file:./security/certs/identity/identity_ca_cert.pem";
     idca.propagate = false;
 
     pkey.name = "dds.sec.auth.private_key";
-    pkey.value = "file:../certs/identity/test_participant_02_private_key.pem";
+    pkey.value = "file:./security/certs/identity/test_participant_02_private_key.pem";
     pkey.propagate = false;
 
     pass.name = "dds.sec.auth.password";
@@ -163,11 +163,11 @@ struct AuthenticationTest : public ::testing::Test
     pass.propagate = false;
 
     idcert.name = "dds.sec.auth.identity_certificate";
-    idcert.value = "file:../certs/identity/test_participant_02_cert.pem";
+    idcert.value = "file:./security/certs/identity/test_participant_02_cert.pem";
     idcert.propagate = false;
 
     cperms.name = "dds.sec.access.permissions";
-    cperms.value = "file:../permissions/permissions_test_participant_02_JoinDomain_signed.p7s";
+    cperms.value = "file:./security/permissions/permissions_test_participant_02_JoinDomain_signed.p7s";
     cperms.propagate = false;
 
     mp2.add_property(idca);
