+++ tests/unit-tests/dds/DCPS/security/UnitTests.mpc
@@ -1,8 +1,26 @@
-project(AccessControlBuiltInImpl): dcpsexe, dcps_test, opendds_security, googlemock {
+project(*AccessControlBuiltInImpl): dcpsexe, dcps_test, opendds_security, googlemock {
   requires += no_opendds_safety_profile
-  exename = AccessControlBuiltInImpl
+  exename = *
 
   Source_Files {
     AccessControlBuiltInImpl.cpp
+  }
+}
+
+project(*AuthenticationBuiltInImpl): dcpsexe, dcps_test, opendds_security, googletest {
+  requires += no_opendds_safety_profile
+  exename = *
+
+  Source_Files {
+    AuthenticationBuiltInImpl.cpp
+  }
+}
+
+project(*CryptoBuiltInImpl): dcpsexe, dcps_test, opendds_security, googletest {
+  requires += no_opendds_safety_profile
+  exename = *
+
+  Source_Files {
+    CryptoBuiltInImpl.cpp
   }
 }
