+++ tests/unit-tests/dds/DCPS/security/framework/UnitTests.mpc
@@ -1,0 +1,9 @@
+project(*SecurityConfig): dcpsexe, dcps_test, googlemock {
+  requires += no_opendds_safety_profile
+  avoids += no_opendds_security
+  exename = *
+
+  Source_Files {
+    SecurityConfig.cpp
+  }
+}
